Implement login and register in accounts context
Remove old implementations
